% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_zzz_package.R
\docType{package}
\name{desfix-package}
\alias{desfix-package}
\alias{desfix}
\title{The 'desfix' package.}
\description{
Demonstrate statistics ideas and concepts through interactive
    Shiny applications.
}
