% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_shiny.R
\name{df_shiny}
\alias{df_shiny}
\title{Run Web-Based application}
\usage{
df_shiny(appname = "boundary", pkgname = "statidea")
}
\description{
Call Shiny to run \code{statidea} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI
}
