% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_simu.R
\name{dose_algorithm_1}
\alias{dose_algorithm_1}
\title{Dose escalation algorithm}
\usage{
dose_algorithm_1(cur_data, lst_design, ...)
}
\description{
0:stay
}
\details{
-12:stop after 6 patients
-13:stop before 6 patients
-14:stop for both DLT and Over FIX
-15:stop for Over FIX
-16:stop for DLT

11: escalate after the 1st patient
12: escalate after 6 patients
13: escalate before 6 patients

2:selected
}
