% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_simu.R
\name{generate_patients}
\alias{generate_patients}
\title{Simulate patients}
\usage{
generate_patients(
  mean_log = c(2.88, 3.8, 4.27, 4.49),
  sd_log = rep(0.47, 4),
  dlt_rate = c(0.01, 0.01, 0.02, 0.02),
  n_patients = 6
)
}
\description{
the lengths of dose_level, mean_log, sd_log and dlt_rate must be the same
}
\details{
right now let n_patients to be the same across dose levels
}
