% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_simu.R
\name{generate_patients_cv}
\alias{generate_patients_cv}
\title{Simulate patients}
\usage{
generate_patients_cv(mean_raw, cv, ...)
}
\arguments{
\item{cv}{Coefficient of variation. If it is scalar, the CV is assumed to be
the same across dose levels. If it is a vector, it has to be the same
length as mean_raw}
}
\description{
Simulate patients
}
